docshell/test/mochitest/test_bug1747033.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test history after loading multipart</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
  <script>
    SimpleTest.waitForExplicitFinish();

    function runTest() {
      let bc = new BroadcastChannel("bug1747033");
      new Promise(resolve => {
        bc.addEventListener("message", ({ data: { historyLength } }) => {
          is(historyLength, 1, "Correct length for first normal load.");

          resolve();
        }, { once: true });

        window.open("file_bug1747033.sjs", "", "noopener");
      }).then(() => {
        return new Promise(resolve => {
          let loaded = 0;
          bc.addEventListener("message", function listener({ data: { historyLength } }) {
            ++loaded;

            is(historyLength, 2, `Correct length for multipart load ${loaded}.`);

            // We want 3 parts in total.
            if (loaded < 3) {
              if (loaded == 2) {
                // We've had 2 parts, make the server send the last part.
                fetch("file_bug1747033.sjs?sendLastPart");
              } else {
                fetch("file_bug1747033.sjs?sendNextPart");
              }
              return;
            }

            bc.removeEventListener("message", listener);
            resolve();
          });

          bc.postMessage({ cmd: "load", arg: "file_bug1747033.sjs?multipart" });
        });
      }).then(() => {
        return new Promise(resolve => {
          bc.addEventListener("message", ({ data: { historyLength } }) => {
            is(historyLength, 2, "Correct length after calling replaceState in multipart.");

            resolve();
          }, { once: true });

          bc.postMessage({ cmd: "replaceState", arg: "file_bug1747033.sjs?replaced" });
        });
      }).then(() => {
        return new Promise(resolve => {
          bc.addEventListener("message", ({ data: { historyLength } }) => {
            is(historyLength, 3, "Correct length for first normal load after multipart.");

            resolve();
          }, { once: true });

          bc.postMessage({ cmd: "load", arg: "file_bug1747033.sjs" });
        });
      }).then(() => {
        return new Promise(resolve => {
          let goneBack = 0;
          bc.addEventListener("message", function listener({ data: { historyLength } }) {
            ++goneBack;

            is(historyLength, 3, "Correct length after going back.");

            if (goneBack == 1) {
              bc.postMessage({ cmd: "back" });
            } else if (goneBack == 2) {
              bc.removeEventListener("message", listener);
              resolve();
            }
          });

          bc.postMessage({ cmd: "back" });
        });
      }).then(() => {
        bc.postMessage({ cmd: "close" });

        SimpleTest.finish();
      });
    }
  </script>
</head>
<body onload="runTest();">
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
</body>
</html>
